deleting photos completely leaving with only one gallery view
