--- frontend/src/app/features/albums/pages/albums-view-photos/albums-view-photos.component.scss
@@ -1,4 +1,0 @@
-.buttons {
-	text-align: right;
-	padding: 5px 0;
-}
